move StretchLengths file to Experimental
